Simplify extern definition and string warnings
