--- src/commands/CMakeLists.txt
@@ -1,2 +1,0 @@
-# Create the commands library
-add_library(commands STATIC tcp_commands.cpp)
