--- src/commands/tcp_commands.cpp
@@ -1,31 +1,0 @@
-/**
- * @file tcp_commands.cpp
- * @author Andrew Obler (andrewobler@gmail.com)
- * @brief Definition of TCP commands and related constants. See include/commands/tcp_commands.hpp.
- * @version 0.1
- * @date 2019-09-20
- * 
- * @copyright Copyright (c) 2019
- * 
- */
-
-#include "commands/tcp_commands.hpp"
-
-char* command_names[] = {
-    "UNSET_VALVE1",
-    "SET_VALVE",
-    "UNSET_VALVE",
-    "SET_VALVE",
-    "UNSET_VALVE",
-    "SET_VALVE",
-    "UNSET_IGNITION",
-    "SET_IGNITION",
-    "SET_WATER",
-    "UNSET_WATER",
-    "SET_GITVC",
-    "UNSET_GITVC",
-    "LEAK_CHECK",
-    "FILL",
-    "FILL_IDLE",
-    "DEF"
-};
